fix compiling

+++ ov_msckf/src/slam_dataset/vi_capture.cpp
@@ -4,7 +4,7 @@
 
 namespace slam_dataset {
 
-void ViCapture::waitStreamingOver(double polling_period_ms) {
+void ViCapture::waitStreamingOver(int polling_period_ms) {
   while (isStreaming()) {
     std::this_thread::sleep_for(
         std::chrono::milliseconds(polling_period_ms));
